agents/rm_agent/pyproject.toml
[project]
name = "rm-agent"
version = "0.1.0"
description = "An AI Agent for RM platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "fastapi>=0.135.2",
    "mcp>=1.26.0",
    "openai>=2.29.0",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
]
